requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.12
anndata==0.8.0
annotated-types==0.6.0
anyio==3.6.2
appdirs==1.4.4
arcp==0.2.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.12.13
asttokens==2.1.0
async-lru==2.0.4
async-timeout==4.0.3
atomicwrites==1.4.1
attmap==0.13.2
attrs==23.2.0
autopage==0.5.1
autopep8==1.6.0
Babel==2.11.0
backcall==0.2.0
bcbio-gff==0.6.9
bcrypt==4.0.1
beautifulsoup4==4.11.1
binaryornot==0.4.4
biom-format==2.1.12
biopython==1.79
black==22.10.0
bleach==5.0.1
boto3==1.24.73
botocore==1.27.73
cattrs==24.1.3
certifi==2022.9.24
cffi==1.15.1
cfgv==3.4.0
chardet==5.1.0
charset-normalizer==2.0.12
checkm-genome==1.2.3
click==8.1.7
click-pathlib==2020.3.13.0
cliff==4.0.0
cloudpickle==2.2.0
cmd2==2.4.2
CMSeq==1.0.4
colorama==0.4.6
coloredlogs==15.0.1
colormath==3.0.0
commonmark==0.9.1
ConfigArgParse==1.5.3
connection-pool==0.0.3
contourpy==1.0.6
cookiecutter==2.1.1
crypt4gh==1.6
cryptography==42.0.4
cycler==0.11.0
Cython==0.29.32
cyvcf2==0.30.18
datrie==0.8.2
dds_cli==2.6.4
debtcollector==2.5.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
DendroPy==4.5.2
depinfo==2.2.0
Deprecated==1.2.18
diff-match-patch==20200713
dill==0.3.6
distlib==0.3.6
docopt==0.6.2
docstring-to-markdown==0.11
docutils==0.19
dogpile.cache==1.1.8
dpath==2.0.6
duplex-tools==0.2.20
edlib==1.3.9
eido==0.2.4
entrypoints==0.4
exceptiongroup==1.2.2
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.0
filetype==1.2.0
flake8==5.0.4
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.4.1
future==0.18.2
gdown==4.5.3
gitdb==4.0.9
GitPython==3.1.29
gtdb-to-taxdump==0.1.9
h11==0.14.0
h5py==3.12.1
hclust2==1.0.0
HTSeq==2.0.2
htsget==0.2.5
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
humanize==4.4.0
humann==3.6
humann2==2.8.2
identify==2.6.10
idna==3.4
imagesize==1.4.1
immutabledict==2.2.1
importlib-metadata==5.0.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.17.0
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
iso8601==1.1.0
isoduration==20.11.0
isort==5.10.1
jaraco.classes==3.2.3
jedi==0.18.1
jeepney==0.8.0
jellyfish==0.9.0
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==1.0.1
joblib==1.2.0
json5==0.9.25
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==7.4.4
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.27.2
jwcrypto==1.5.6
kaleido==0.2.1
kb-python==0.27.3
keyring==23.11.0
keystoneauth1==5.0.0
kiwisolver==1.4.4
lazy-object-proxy==1.8.0
lib_pod5==0.3.23
libhfst-swig==3.16.0b0
linkify-it-py==2.0.3
llvmlite==0.39.1
logmuse==0.2.8
loompy==3.0.7
lzstring==1.0.4
mappy==2.24
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
MetaPhlAn==4.0.3
mistune==2.0.4
more-itertools==9.0.0
msgpack==1.0.4
multidict==6.0.4
multimethod==1.10
multiqc==1.13
munch==2.5.0
mypy-extensions==0.4.3
nanoget==1.18.1
nanomath==1.2.1
NanoPlot==1.41.0
natsort==8.2.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
netaddr==0.8.0
netifaces==0.11.0
networkx==2.8.8
nf-core==3.2.0
ngs-tools==1.8.3
nodeenv==1.9.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.4
numpy-groupies==0.9.20
numpydoc==1.5.0
openstacksdk==0.102.0
os-service-types==1.7.0
osc-lib==2.6.2
oslo.config==9.0.0
oslo.i18n==5.1.0
oslo.serialization==5.0.0
oslo.utils==6.0.1
overrides==7.7.0
oyaml==1.0
packaging==24.1
pandas==2.2.3
pandera==0.18.3
pandocfilters==1.5.0
parasail==1.3.3
parso==0.8.3
pathspec==0.10.2
patsy==0.5.3
pbr==5.11.0
pdiff==1.1.4
pephubclient==0.4.5
peppy==0.40.7
pexpect==4.8.0
PhyloPhlAn==3.0.3
pickleshare==0.7.5
Pillow==9.3.0
piper==0.14.4
pipestat==0.12.1
plac==1.3.5
platformdirs==2.5.2
plotly==5.11.0
pluggy==1.0.0
pod5==0.3.23
polars==0.20.31
pre_commit==4.2.0
prettytable==3.4.1
prometheus-client==0.15.0
prompt-toolkit==3.0.31
psutil==5.6.6
ptyprocess==0.7.0
PuLP==2.6.0
pure-eval==0.2.2
pyarrow==18.0.0
pycodestyle==2.9.1
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydocstyle==6.1.1
pyega3==5.0.2
pyfaidx==0.8.1.3
pyfastx==0.9.1
pyflakes==2.5.0
PyGithub==2.6.1
Pygments==2.13.0
PyJWT==2.10.1
pylint==2.15.8
pylint-venv==2.3.0
pyls-spyder==0.4.0
PyNaCl==1.5.0
pynndescent==0.5.8
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyrsistent==0.19.2
pysam==0.20.0
PySocks==1.7.1
python-cinderclient==9.1.0
python-dateutil==2.8.2
Python-Deprecated==1.1.0
python-json-logger==2.0.7
python-keystoneclient==5.0.1
python-lsp-black==1.2.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.6.0
python-magic==0.4.27
python-novaclient==18.1.0
python-openstackclient==6.0.0
python-slugify==7.0.0
python-swiftclient==4.1.0
pytoolconfig==1.2.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.28
PyYAML==6.0.1
pyzmq==24.0.1
QDarkStyle==3.0.3
qstylizer==0.2.2
QtAwesome==1.2.1
qtconsole==5.4.0
QtPy==2.2.1
questionary==2.1.0
ratelimiter==1.2.0.post0
referencing==0.35.1
refgenconf==0.12.2
refgenie==0.12.1
repo2rocrate==0.1.2
reportlab==3.6.12
requests==2.31.0
requests-cache==1.2.1
requestsexceptions==1.4.0
reretry==0.11.1
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==14.0.0
rich-click==1.8.8
rocrate==0.13.0
rope==1.5.1
rpds-py==0.18.1
Rtree==1.0.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3cmd==2.3.0
s3transfer==0.6.2
scanpy==1.9.1
scikit-learn==1.1.3
scipy==1.9.3
scvelo==0.2.5
sd-lock-utility @ file:///scratch/project_2000828/kkmattil/sd-lock-util
seaborn==0.12.1
SecretStorage==3.3.3
Send2Trash==1.8.3
session-info==1.0.0
shellingham==1.5.4
shortuuid==1.0.11
simplejson==3.17.6
six==1.16.0
sklearn==0.0.post1
smart-open==6.2.0
smmap==5.0.0
snakemake==7.18.1
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spectra==0.0.11
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.4.0
spyder-kernels==2.4.0
stack-data==0.6.0
statsmodels==0.13.5
stdlib-list==0.8.0
stevedore==4.1.0
stopit==1.1.2
tabulate==0.9.0
taxopy==0.12.0
taxpasta==0.6.1
tenacity==8.1.0
terminado==0.17.0
text-unidecode==1.3
textdistance==4.5.0
textual==0.71.0
threadpoolctl==3.1.0
three-merge==0.1.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
toposort==1.7
tornado==6.2
tqdm==4.65.0
traitlets==5.14.3
trogon==0.6.0
typeguard==4.2.1
typer==0.12.0
typer-cli==0.12.0
typer-slim==0.12.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==4.2
ubiquerg==0.8.1
uc-micro-py==1.0.3
ujson==5.6.0
umap-learn==0.5.3
uri-template==1.3.0
url-normalize==2.2.0
urllib3==1.26.7
vbz-h5py-plugin==1.0.1
velocyto==0.17.17
virtualenv==20.16.6
watchdog==2.2.0
wcwidth==0.2.5
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
whatthepatch==1.0.3
widgetsnbextension==4.0.3
wrapt==1.14.1
wurlitzer==3.0.3
yacman==0.9.3
yapf==0.32.0
yarl==1.9.4
yte==1.5.1
zipp==3.10.0
zstandard==0.19.0
